Handle maps where 0x0 is a valid key
With OS-X / ADB keycodes 0x0 refers to key 'A', so cannot be
used as the default in the array. We must use 0xff instead and
thus explicitly fill every array entry with 0xff if no mapping
is available.

To handle this for enum based source maps, we need to use the
magic syntax

   [0 .... max] = value,

to set the default value in all elements, and then override
it afterwards for mappings which are known.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>