keymap-gen: Name sections in pod output
Properly formatted man pages have at least a NAME section,
which tools like apropos(1) parse and index.

The current implementation of PodLanguageGenerator doesn't
generate this section, and even if the user were to try and
work around the limitation by passing --varname=NAME, there
would still be nothing in the section to actually parse.

Change the generator so that varname will be used as the
contents of the NAME section, add a DESCRIPTION section in
which values are listed.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>