Add .gitattributes file
The files marked as export-ignore here are not going to be
included in the tarball produced by 'meson dist' when using
meson >= 0.60. Older versions of meson excluded some files
automatically, but that no longer happens.

Note that, while keycodemapdb itself doesn't use meson or
really produce release tarballs, the presence of this file
will help meson-using projects that include it as a git
submodule - for example libvirt.

Signed-off-by: Andrea Bolognani <abologna@redhat.com>