Fix macOS "ISO" key
This is actually the common ISO "102nd" key, found between the left
shift and "Z".

Note that this is still technically wrong, as macOS swaps the key codes
for ANSI_Grave and ISO_Section for ISO layouts. This magic is not
something we can encode here though, so that swapping will have to be
done by the implementations using this data.